move model_validator to tests
